requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
gunicorn==20.1.0
Pillow==10.0.0
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0